refactor(mobileapp): rename uniques to single for the single page folder
